+++ src/onegov.gazette/HISTORY.rst
@@ -4,6 +4,9 @@ Changelog
 - Updates translation.
   [msom]
 
+- Uses a bigger reject comment field.
+  [msom]
+
 1.9.6 (2017-11-20)
 ~~~~~~~~~~~~~~~~~~~